backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==3.0.2
Flask-RESTful==0.3.10
gotrue==2.4.1
postgrest==0.16.1
python-dotenv==1.0.1
supabase==2.4.0
Werkzeug==3.0.1
